move fabric datasource out of experimental dir
